chore: moved ConnectionInfoProvider and the locators to @mongodb-js/compass-connections
