+++ packages/compass-connections/src/connection-scoped-app-registry.ts
@@ -6,7 +6,7 @@ import {
 import {
   type ConnectionInfoAccess,
   connectionInfoAccessLocator,
-} from '@mongodb-js/connection-storage/provider';
+} from './connection-info-provider';
 
 export type ConnectionScopedAppRegistryLocator<
   T extends string,
