+++ packages/compass-indexes/src/components/indexes/indexes.tsx
@@ -25,7 +25,7 @@ import {
 } from '../search-indexes-modals';
 import type { IndexView } from '../../modules/index-view';
 import { usePreference } from 'compass-preferences-model/provider';
-import { useConnectionInfo } from '@mongodb-js/connection-storage/provider';
+import { useConnectionInfo } from '@mongodb-js/compass-connections/provider';
 import { getAtlasSearchIndexesLink } from '../../utils/atlas-search-indexes-link';
 
 // This constant is used as a trigger to show an insight whenever number of
